renamed the util package to component
