+++ src/main/java/com/bootcamp/demo/model/component/BatteryStatus.java
@@ -1,4 +1,4 @@
-package com.bootcamp.demo.model.util;
+package com.bootcamp.demo.model.component;
 
 public enum BatteryStatus {
     LOW("LOW"),
